Separated the Javascript code into multiple files
